chore: revert #5730
