+++ generators/dart/loops.js
@@ -152,7 +152,7 @@ Blockly.Dart['controls_flow_statements'] = function(block) {
     xfix += Blockly.Dart.injectId(Blockly.Dart.STATEMENT_SUFFIX, block);
   }
   if (Blockly.Dart.STATEMENT_PREFIX) {
-    const loop = Blockly.loopMixin
+    const loop = Blockly.Constants.Loops
         .CONTROL_FLOW_IN_LOOP_CHECK_MIXIN.getSurroundLoop(block);
     if (loop && !loop.suppressPrefixSuffix) {
       // Inject loop's statement prefix here since the regular one at the end
