+++ generators/javascript/loops.js
@@ -12,8 +12,8 @@
 
 goog.provide('Blockly.JavaScript.loops');
 
+goog.require('Blockly.Constants.Loops');
 goog.require('Blockly.JavaScript');
-goog.require('Blockly.loopMixin');
 goog.require('Blockly.utils.string');
 
 
@@ -167,7 +167,7 @@ Blockly.JavaScript['controls_flow_statements'] = function(block) {
         block);
   }
   if (Blockly.JavaScript.STATEMENT_PREFIX) {
-    const loop = Blockly.loopMixin
+    const loop = Blockly.Constants.Loops
         .CONTROL_FLOW_IN_LOOP_CHECK_MIXIN.getSurroundLoop(block);
     if (loop && !loop.suppressPrefixSuffix) {
       // Inject loop's statement prefix here since the regular one at the end
