+++ generators/lua/loops.js
@@ -156,7 +156,7 @@ Blockly.Lua['controls_flow_statements'] = function(block) {
     xfix += Blockly.Lua.injectId(Blockly.Lua.STATEMENT_SUFFIX, block);
   }
   if (Blockly.Lua.STATEMENT_PREFIX) {
-    const loop = Blockly.loopMixin
+    const loop = Blockly.Constants.Loops
         .CONTROL_FLOW_IN_LOOP_CHECK_MIXIN.getSurroundLoop(block);
     if (loop && !loop.suppressPrefixSuffix) {
       // Inject loop's statement prefix here since the regular one at the end
