+++ generators/php/loops.js
@@ -153,7 +153,7 @@ Blockly.PHP['controls_flow_statements'] = function(block) {
     xfix += Blockly.PHP.injectId(Blockly.PHP.STATEMENT_SUFFIX, block);
   }
   if (Blockly.PHP.STATEMENT_PREFIX) {
-    const loop = Blockly.loopMixin
+    const loop = Blockly.Constants.Loops
         .CONTROL_FLOW_IN_LOOP_CHECK_MIXIN.getSurroundLoop(block);
     if (loop && !loop.suppressPrefixSuffix) {
       // Inject loop's statement prefix here since the regular one at the end
