+++ scripts/gulpfiles/chunks.json
@@ -1,7 +1,7 @@
 {
   "chunk": [
     "requires:259",
-    "all:10:requires",
+    "all:9:requires",
     "all1:11:requires",
     "all2:11:requires",
     "all3:11:requires",
@@ -11,64 +11,63 @@
   "js": [
     "./core/serialization/workspaces.js",
     "./core/serialization/variables.js",
-    "./core/widgetdiv.js",
-    "./core/clipboard.js",
-    "./core/menuitem.js",
-    "./core/menu.js",
-    "./core/contextmenu.js",
-    "./core/contextmenu_items.js",
-    "./core/shortcut_items.js",
-    "./core/theme/zelos.js",
-    "./core/theme/themes.js",
-    "./core/renderers/geras/constants.js",
-    "./core/renderers/geras/highlight_constants.js",
-    "./core/renderers/geras/measurables/inline_input.js",
-    "./core/renderers/geras/path_object.js",
-    "./core/renderers/geras/renderer.js",
-    "./core/renderers/geras/measurables/statement_input.js",
-    "./core/renderers/geras/info.js",
-    "./core/renderers/geras/highlighter.js",
-    "./core/renderers/geras/drawer.js",
-    "./core/renderers/geras/geras.js",
-    "./core/loop_mixin.js",
-    "./core/renderers/minimalist/constants.js",
-    "./core/renderers/minimalist/renderer.js",
-    "./core/renderers/minimalist/info.js",
-    "./core/renderers/minimalist/drawer.js",
-    "./core/renderers/minimalist/minimalist.js",
+    "./core/inject.js",
+    "./core/flyout_vertical.js",
+    "./core/toolbox/toolbox.js",
+    "./core/interfaces/i_styleable.js",
+    "./core/flyout_horizontal.js",
+    "./core/flyout_base.js",
+    "./core/flyout_metrics_manager.js",
+    "./core/field_variable.js",
+    "./core/field_number.js",
+    "./core/field_multilineinput.js",
+    "./core/field_label_serializable.js",
+    "./core/field_colour.js",
+    "./core/field_checkbox.js",
+    "./core/field_angle.js",
+    "./core/toolbox/collapsible_category.js",
+    "./core/renderers/zelos/measurables/top_row.js",
+    "./core/renderers/zelos/measurables/inputs.js",
+    "./core/renderers/zelos/measurables/row_elements.js",
+    "./core/renderers/zelos/marker_svg.js",
+    "./core/renderers/zelos/renderer.js",
+    "./core/field_textinput.js",
+    "./core/field_image.js",
+    "./core/renderers/zelos/info.js",
+    "./core/renderers/zelos/path_object.js",
+    "./core/renderers/zelos/drawer.js",
+    "./core/renderers/zelos/constants.js",
+    "./core/renderers/zelos/measurables/bottom_row.js",
+    "./core/renderers/zelos/zelos.js",
     "./core/renderers/thrasos/renderer.js",
     "./core/renderers/thrasos/info.js",
     "./core/renderers/thrasos/thrasos.js",
-    "./core/renderers/zelos/measurables/bottom_row.js",
-    "./core/renderers/zelos/constants.js",
-    "./core/renderers/zelos/path_object.js",
-    "./core/field_image.js",
-    "./core/field_textinput.js",
-    "./core/renderers/zelos/marker_svg.js",
-    "./core/renderers/zelos/renderer.js",
-    "./core/renderers/zelos/measurables/row_elements.js",
-    "./core/renderers/zelos/measurables/inputs.js",
-    "./core/renderers/zelos/measurables/top_row.js",
-    "./core/renderers/zelos/info.js",
-    "./core/renderers/zelos/drawer.js",
-    "./core/renderers/zelos/zelos.js",
-    "./core/toolbox/collapsible_category.js",
-    "./core/field_angle.js",
-    "./core/field_checkbox.js",
-    "./core/field_colour.js",
-    "./core/field_label_serializable.js",
-    "./core/field_multilineinput.js",
-    "./core/field_number.js",
-    "./core/field_variable.js",
-    "./core/flyout_metrics_manager.js",
-    "./core/flyout_base.js",
-    "./core/generator.js",
-    "./core/flyout_horizontal.js",
-    "./core/interfaces/i_styleable.js",
-    "./core/toolbox/toolbox.js",
-    "./core/flyout_vertical.js",
-    "./core/inject.js",
+    "./core/renderers/minimalist/renderer.js",
+    "./core/renderers/minimalist/info.js",
+    "./core/renderers/minimalist/drawer.js",
+    "./core/renderers/minimalist/constants.js",
+    "./core/renderers/minimalist/minimalist.js",
+    "./core/renderers/geras/measurables/statement_input.js",
+    "./core/renderers/geras/path_object.js",
+    "./core/renderers/geras/renderer.js",
+    "./core/renderers/geras/measurables/inline_input.js",
+    "./core/renderers/geras/info.js",
+    "./core/renderers/geras/highlight_constants.js",
+    "./core/renderers/geras/highlighter.js",
+    "./core/renderers/geras/drawer.js",
+    "./core/renderers/geras/constants.js",
+    "./core/renderers/geras/geras.js",
+    "./core/theme/zelos.js",
+    "./core/theme/themes.js",
+    "./core/shortcut_items.js",
+    "./core/contextmenu_items.js",
+    "./core/menu.js",
+    "./core/menuitem.js",
+    "./core/clipboard.js",
+    "./core/widgetdiv.js",
+    "./core/contextmenu.js",
     "./core/blockly.js",
+    "./core/generator.js",
     "./core/field_dropdown.js",
     "./core/events/workspace_events.js",
     "./core/events/events_toolbox_item_select.js",
@@ -267,13 +266,13 @@
     "./core/utils/global.js",
     "./core/msg.js",
     "./closure/goog/base_minimal.js",
+    "./blocks/loops.js",
     "./core/requires.js",
     "./blocks/variables_dynamic.js",
     "./blocks/variables.js",
     "./blocks/text.js",
     "./blocks/procedures.js",
     "./blocks/math.js",
-    "./blocks/loops.js",
     "./blocks/logic.js",
     "./blocks/lists.js",
     "./blocks/colour.js",
