+++ scripts/migration/renamings.js
@@ -254,11 +254,6 @@ const renamings = {
             {module: 'Blockly.Extensions', export: 'runAfterPageLoad'},
       },
     },
-    'Blockly.Constants.Loops': {
-      exports: {
-        CONTROL_FLOW_IN_LOOP_CHECK_MIXIN: {module: 'Blockly.loopMixin'},
-      },
-    },
     'Blockly.Events.Abstract': {
       export: 'Abstract',
       path: 'Blockly.Events.Abstract',
